Separate integration tests a bit.
